feat(tasks)!: update the output and export to SW spec
BREAKING CHANGE: This has changed the output/export to what Serverless
Workflow intended. This is pretty poorly documented, so I misunderstood
it at first.

The basic idea is that the last task defaults it's own output as the
response - eg, if you have a `set` task, it returns that or, if you have
a `callHttp` task, it returns the result.

If you want to combine things, you need to `export` it to the `$context`
and then merge it

```yaml
do:
  - data:
      set:
        envvar: ${ $env.EXAMPLE_ENVVAR }
        uuid: ${ uuid }
```

This will output `{ "envvar": "<envvar value>", "uuid": "<some uuid>" }`

```yaml
do:
  - data:
      set:
        envvar: ${ $env.EXAMPLE_ENVVAR }
        uuid: ${ uuid }
  - data1:
      set:
        hello: world
```

This will output `{ "hello": "world" }`

```yaml
do:
  - data:
      set:
        envvar: ${ $env.EXAMPLE_ENVVAR }
        uuid: ${ uuid }
      export:
        # This sets the data to `$context.data`
        as:
          data: ${ . }
  - data1:
      set:
        hello: world
      output:
        # This merges all the data in the `$context` with the data set here
        as: '${ $context + { "data1": . } }'
```

This will output:

```json
{
  "data": {
    "envvar": "<envvar value>",
    "uuid": "<some uuid>"
  },
  "data1": {
    "hello": "world"
  }
}
```

Signed-off-by: Simon Emms <simon@simonemms.com>